migrate qmu, generate_asimov_data, loglambdav, and hypotest
 - update docs api
 - fix up api a bit better
