+++ docs/api.rst
@@ -109,6 +109,19 @@ Interpolators
    code4
    code4p
 
+Statistics
+----------
+
+.. currentmodule:: pyhf.stats
+
+.. autosummary::
+   :toctree: _generated/
+
+   test_statistics.hypotest
+   test_statistics.qmu
+   utils.loglambdav
+   utils.generate_asimov_data
+
 Exceptions
 ----------
 
@@ -141,9 +154,5 @@ Utilities
 .. autosummary::
    :toctree: _generated/
 
-   generate_asimov_data
-   loglambdav
    pvals_from_teststat
    pvals_from_teststat_expected
-   qmu
-   hypotest
